add AI

+++ model/src/main/java/com/thm/aiarena/model/AResource.java
@@ -2,5 +2,5 @@ package com.thm.aiarena.model;
 
 public interface AResource {
     int getType();
-    long getAmount();
+    int getAmount();
 }
